move spec test file in directory 'spec'
